+++ lib/overcommit/hook_context/pre_commit.rb
@@ -14,7 +14,7 @@ module Overcommit::HookContext
       store_merge_state
       store_cherry_pick_state
 
-      if any_changes?
+      if !initial_commit? && any_changes?
         @changes_stashed = true
         `git stash save --keep-index --quiet #{<<-MSG}`
           "Overcommit: Stash of repo state before hook run at #{Time.now}"
@@ -28,7 +28,9 @@ module Overcommit::HookContext
     # Restore unstaged changes and reset file modification times so it appears
     # as if nothing ever changed.
     def cleanup_environment
-      `git reset --hard` # Ensure working tree is clean before popping stash
+      unless initial_commit?
+        `git reset --hard` # Ensure working tree is clean before popping stash
+      end
 
       if @changes_stashed
         `git stash apply --index --quiet`
@@ -43,8 +45,8 @@ module Overcommit::HookContext
     # Renames and deletions are ignored, since there should be nothing to check.
     def modified_files
       @modified_files ||=
-        `git diff --cached --name-only --diff-filter=ACM --ignore-submodules=all`.
-          split("\n").
+        `git diff --cached --name-only -z --diff-filter=ACM --ignore-submodules=all`.
+          split("\0").
           map { |relative_file| File.expand_path(relative_file) }
     end
 
@@ -57,10 +59,20 @@ module Overcommit::HookContext
 
   private
 
-    # Returns whether there are any changes between the working tree or index
-    # and HEAD.
+    # Returns whether there are any changes to the working tree, staged or
+    # otherwise.
     def any_changes?
-      !`git diff --name-only --ignore-submodules=all HEAD`.strip.empty?
+      modified_files = `git status -z --untracked-files=no`.
+        split("\0").
+        map { |line| line.gsub(/[^\s]+\s+(.+)/, '\\1') }
+
+      modified_files.any?
+    end
+
+    # Returns whether the current git branch is empty (has no commits).
+    def initial_commit?
+      return @initial_commit unless @initial_commit.nil?
+      @initial_commit = !Overcommit::Utils.execute(%w[git rev-parse HEAD]).success?
     end
 
     DIFF_HUNK_REGEX = /