+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Prevent `JsonSyntax` pre-commit hook from failing if `json_class` key
   is present in JSON
 * Prevent `HardTabs` pre-commit hook from warning on tabs in Makefiles
+* Fix bug where `overcommit` hooks would fail for initial commit to repo
 
 ## 0.12.0